git state report config
